Make gem-src as a rbenv plugin
By making use of the rbenv's hook feature, a load path can be injected to all
rubies under rbenv. This makes gem-src enabled for all gem and bundle commands
invoked through rbenv.

This is effective only if it is installed as a rbenv plugin like described in
README, so it does not disturb the existing users using this as a gem plugin.